Lista 'Por hacer' actualizada. Casteos reemplazados. Subdivisión de escenas terminada. Faltan los power-ups y etc.

+++ Pong/src/Ball.cpp
@@ -8,8 +8,8 @@ Ball ball;
 
 void InitBall(Ball &ball)
 {
-	ball.position = { (float)screenWidth / 2, (float)screenHeight / 2 };
-	ball.direction.x = (float)GetRandomValue(minBallSpeed, maxBallSpeed - 1); ball.direction.y = (float)(maxBallSpeed - ball.direction.x);
+	ball.position = { screenWidth / 2, screenHeight / 2 };
+	ball.direction.x = static_cast<float>(GetRandomValue(static_cast<int>(minBallSpeed), static_cast<int>(maxBallSpeed - 1))); ball.direction.y = maxBallSpeed - ball.direction.x;
 	ball.color = RAYWHITE;
 
 	ball.radius = 15;
